jwt-extended-changes

+++ resources/item.py
@@ -1,12 +1,27 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (jwt_required,
+                                get_jwt_claims,
+                                jwt_optional,
+                                get_jwt_identity,
+                                fresh_jwt_required)
 from models.item import ItemModel
 
 
 class Items(Resource):
-    @jwt_required()
+    @jwt_optional
     def get(self):
-        return {'items':[item.json() for item in ItemModel.query.all()]}  #list(map(lambda x:x.json() ,ItemModel.query.all()
+        user_id=get_jwt_identity()
+        items=[item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items':items},200
+
+        else:
+            return {
+                'items':[item['name'] for item in items],
+                'message':'More information available if logged in'
+            },200
+
+        return {'items':[item.json() for item in ItemModel.find_all()]}  #list(map(lambda x:x.json() ,ItemModel.query.all()
 
 class Item(Resource):
     parser = reqparse.RequestParser()
@@ -20,7 +35,7 @@ class Item(Resource):
                         help='Every item needs a store id')
 
 
-    @jwt_required()
+    @jwt_required
     def get(self,name):
         item=ItemModel.find_by_name(name)
         if item:
@@ -29,7 +44,7 @@ class Item(Resource):
             return {'message':'item not found'} ,404
 
 
-
+    @fresh_jwt_required
     def post(self,name):
         if ItemModel.find_by_name(name):
             return {'message':'Item with the given name already exists'},400
@@ -46,7 +61,11 @@ class Item(Resource):
         return item.json(), 201
 
 
+    @jwt_required
     def delete(self,name):
+        claims=get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message':'Requires admin previlege'},401
         item=ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
